requirements.txt
requests>=2.31.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
gspread>=5.12.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0